app/requirements.txt
Package             Version
------------------- ---------
asgiref             3.6.0
attrs               22.2.0
certifi             2022.12.7
charset-normalizer  3.0.1
Django              4.1.5
django-cors-headers 3.13.0
djangorestframework 3.12.4
drf-spectacular     0.15.1
idna                3.4
inflection          0.5.1
jsonschema          4.17.3
pip                 22.3.1
pyrsistent          0.19.3
pytz                2022.7.1
PyYAML              6.0
requests            2.28.2
setuptools          65.6.3
sqlparse            0.4.3
uritemplate         4.1.1
urllib3             1.26.14
